:recycle: Move processingSpeed to FormBtc
